chore: update extension host log message
