+++ packages/core-node/src/bootstrap/app.ts
@@ -78,7 +78,6 @@ export class ServerApp implements IServerApp {
       blockPatterns: opts.blockPatterns,
       extHostIPCSockPath: opts.extHostIPCSockPath,
       extHostForkOptions: opts.extHostForkOptions,
-      killExtensionHostProcessWhenDisconnected: opts.killExtensionHostProcessWhenDisconnected ?? true,
     };
     this.bindProcessHandler();
     this.initBaseProvider();
