+++ packages/extension/src/common/const.ts
@@ -1,0 +1,1 @@
+export const CONNECTION_HANDLE_BETWEEN_EXTENSION_AND_MAIN_THREAD = 'ExtMainThreadConnection';
