+++ packages/extension/src/node/extension.host.manager.ts
@@ -1,4 +1,3 @@
-import assert from 'assert';
 import cp from 'child_process';
 
 import isRunning from 'is-running';
@@ -18,19 +17,24 @@ export class ExtensionHostManager implements IExtensionHostManager {
   }
 
   init() {
-    // 这个 debug 插件 https://github.com/microsoft/vscode/blob/1.44.2/extensions/debug-auto-launch/src/extension.ts#L120
-    // 依赖了 vscode pid 来查找子进程
+    // Debug Auto Launch 插件 https://github.com/microsoft/vscode/blob/1.44.2/extensions/debug-auto-launch/src/extension.ts#L120
+    // 依赖了 VSCODE_PID 来查找子进程
     process.env['VSCODE_PID'] = String(process.pid);
   }
+
   fork(modulePath: string, ...args: any[]) {
     const extProcess = cp.fork(modulePath, ...args);
     this.processMap.set(extProcess.pid, extProcess);
     return extProcess.pid;
   }
+
   send(pid: number, message: string) {
     return new Promise<void>((resolve, reject) => {
       const extProcess = this.processMap.get(pid);
-      assert(extProcess, "[ExtensionHostManager] Can't find process with pid: " + pid);
+      if (!extProcess) {
+        reject(new Error(`Can't find process with pid ${pid}`));
+        return;
+      }
       extProcess.send(message, (err) => {
         if (err) {
           reject(err);
@@ -40,9 +44,11 @@ export class ExtensionHostManager implements IExtensionHostManager {
       });
     });
   }
+
   isRunning(pid: number) {
     return isRunning(pid);
   }
+
   treeKill(pid: number) {
     return new Promise<void>((resolve, reject) => {
       treeKill(pid, (err) => {
@@ -54,11 +60,12 @@ export class ExtensionHostManager implements IExtensionHostManager {
       });
     });
   }
+
   kill(pid: number, signal?: NodeJS.Signals) {
     const extProcess = this.processMap.get(pid);
-    assert(extProcess);
-    extProcess!.kill(signal);
+    extProcess?.kill(signal);
   }
+
   isKilled(pid: number) {
     const extProcess = this.processMap.get(pid);
     if (!extProcess) {
@@ -73,11 +80,13 @@ export class ExtensionHostManager implements IExtensionHostManager {
 
   onOutput(pid: number, listener: (output: Output) => void) {
     const extProcess = this.processMap.get(pid);
-    assert(extProcess);
-    extProcess!.stdout!.setEncoding('utf8');
-    extProcess!.stderr!.setEncoding('utf8');
-    const onStdout = Event.fromNodeEventEmitter<string>(extProcess!.stdout!, 'data');
-    const onStderr = Event.fromNodeEventEmitter<string>(extProcess!.stderr!, 'data');
+    if (!extProcess) {
+      return;
+    }
+    extProcess.stdout.setEncoding('utf8');
+    extProcess.stderr.setEncoding('utf8');
+    const onStdout = Event.fromNodeEventEmitter<string>(extProcess.stdout, 'data');
+    const onStderr = Event.fromNodeEventEmitter<string>(extProcess.stderr, 'data');
     const onOutput = Event.any(
       Event.map(onStdout, (o) => ({ type: OutputType.STDOUT, data: `%c${o}`, format: [''] })),
       Event.map(onStderr, (o) => ({ type: OutputType.STDERR, data: `%c${o}`, format: ['color: red'] })),
@@ -102,14 +111,18 @@ export class ExtensionHostManager implements IExtensionHostManager {
 
   onExit(pid: number, listener: (code: number, signal: string) => void) {
     const extProcess = this.processMap.get(pid);
-    assert(extProcess);
-    extProcess!.once('exit', listener);
+    if (!extProcess) {
+      return;
+    }
+    extProcess.once('exit', listener);
   }
 
   onMessage(pid: number, listener: (msg: any) => void): MaybePromise<void> {
     const extProcess = this.processMap.get(pid);
-    assert(extProcess);
-    extProcess!.on('message', listener);
+    if (!extProcess) {
+      return;
+    }
+    extProcess.on('message', listener);
   }
 
   disposeProcess(pid: number) {
